requirements.txt
# Production dependencies (used in Dockerfile)
fastapi>=0.115
uvicorn[standard]>=0.34
httpx[socks]>=0.28
pycryptodome>=3.21
pydantic>=2.10
pydantic-settings>=2.7
psycopg2-binary>=2.9
aiofiles>=24.1
click>=8.1
apscheduler>=3.10,<4
pandas>=2.2
google-cloud-storage>=2.18
google-auth>=2.27,<2.35  # >=2.35 uses HTTPS metadata endpoint, breaks on GCE without mTLS
pyjwt>=2.8
bcrypt>=4.0
email-validator>=2.1
slowapi>=0.1.9
cachetools>=5.3
